Version 1.14
  * More cross-platform services: deluged, deluge-web, multilog@, sissylog@, openvpn@, wpa_supplicant, polkitd, timedated, localed, hostnamed, NetworkManager, ModemManager, sickbeard, plexmediaserver, ossec@dbd, ossec@agentlessd, ossec@csyslogd, ossec@execd, ossec@agentd, ossec@logcollector, ossec@syscheckd, ossec@maild, ossec@analysisd, ossec@remoted, ossec@monitord
  * The cups.service service has been renamed org.cups.cups.service to follow the service names used by CUPS packaging.
  * Fixed a bug in foreground and background where nested separators on the right-hand-side were not being preserved.
  * The previous regular sockets, sysinit services, and standard targets packages are now all merged into the bundles package.
  * There is now a new user VT package that runs new-style user VT services.
  * There is now a new kernel VT package that runs old-style kernel VT services.
  * The nosh-systemd-services package now automatically runs various nosh subsystems under systemd.
  * The maintainer scripts now create the cyclog@ user accounts, convert any systemd presets, and preset the services.
  * Fixed a bug in service-is-{ok,up} where a non-int exception was being thrown.
  * Fixed a bug in convert-systemd-units where it wasn't handling quoted whitespace in settings that take lists.
  * openstack-swift@.service is now swift@.service to make the usernames shorter.

+++ source/service-is-ok.cpp
@@ -1,4 +1,4 @@
-/* COPYING ******************************************************************
+/* coPYING ******************************************************************
 For copyright and licensing terms, see the file named COPYING.
 // **************************************************************************
 */
@@ -46,11 +46,11 @@ service_is_ok (
 	}
 	const char * name(args[0]);
 	const int dir_fd(open_dir_at(AT_FDCWD, name));
-	if (0 > dir_fd) throw EXIT_TEMPORARY_FAILURE;
+	if (0 > dir_fd) throw static_cast<int>(EXIT_TEMPORARY_FAILURE);
 	const int ok_fd(open_writeexisting_at(dir_fd, "ok"));
 	if (0 <= ok_fd) throw EXIT_SUCCESS;
 	const int supervise_ok_fd(open_writeexisting_at(dir_fd, "supervise/ok"));
 	if (0 <= supervise_ok_fd) throw EXIT_SUCCESS;
 
-	throw EXIT_PERMANENT_FAILURE;
+	throw static_cast<int>(EXIT_PERMANENT_FAILURE);
 }
